Defaulting to latest
The issue with having this set to the latest tag is those build of master SHOULD get master/latest else you can end up with wierd edge cases due to the fact the tag is carried through to things like the images versions and the github release. This change should make it work as as expected i.e. if you build the binaries off master you get latest images and if you pull from a release it's tagged with an actual release version