Build of the local users branch #1181 so that needs merging first. This changes on top of this are
acf862b
2ea2a83
The PR addresses the issues of when you have drift on the official helm charts. At the moment when you build it takes the tag (which is last tag on branch). One, the tags might not have been published with assets so it breaks and there's no helm chart to download and two, the last tag can be old and drifted from the master. Essentially we need everything the binary rely's on to be snapshot into the binary to ensure everything is at the right version. This PR thus bundles the helm into the binary ensuring everything matches up.

The issue cropped up when some people tried Kore and hit issues building off master with assets pinned at different points.

Note: we should merge #1188 beforehand as it fixes issues when build off master.