Add new ServiceCatalog CRD and relevant API endpoints
A ServiceCatalog refers to a Helm repository which contains one or more Helm charts.

All Helm Chart names will be created as a ServiceKind and all Helm Chart versions will created as ServicePlans.

These services will be deployed as a Helm application.

A service can be a system service (annotation: kore.appvia.io/system=true), in which case it will be deployed into the Kore cluster, or it can be deployed into any cluster otherwise.